approach1/requirements.txt
torch
transformers
datasets
pytesseract
opencv-python
Pillow
pandas
openpyxl
streamlit
